refactor: migrates loadContracts to query, creates lib for methods
